Rename connected account to account
Rename connected account slice to just "account"
