+++ src/hooks/useSaveConnectedAddressToStore.ts
@@ -1,7 +1,7 @@
 import { useWeb3React } from "@web3-react/core"
 import { useEffect } from "react"
 import { useDispatch } from "react-redux"
-import { setConnectedAccountAddress } from "../store/connected-account"
+import { setAccountAddress } from "../store/account"
 
 export const useSaveConnectedAddressToStore = () => {
   const { account } = useWeb3React()
@@ -9,6 +9,6 @@ export const useSaveConnectedAddressToStore = () => {
 
   useEffect(() => {
     const address = account ? account : ""
-    dispatch(setConnectedAccountAddress(address))
+    dispatch(setAccountAddress(address))
   }, [account])
 }
