--- src/store/connected-account/index.ts
@@ -1,1 +1,0 @@
-export * from "./connectedAccountSlice"
